@@ -7,27 +7,27 @@ environment:
   sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
-  bloc: ^7.2.1
+  bloc: ^8.1.0
   equatable: ^2.0.3
-  extension_google_sign_in_as_googleapis_auth: ^2.0.4
+  extension_google_sign_in_as_googleapis_auth: ^2.0.7
   flutter:
     sdk: flutter
-  flutter_bloc: ^7.3.3
+  flutter_bloc: ^8.1.1
   flutter_localizations:
     sdk: flutter
-  google_sign_in: ^5.2.1
-  googleapis: ^7.0.0
+  google_sign_in: ^5.4.2
+  googleapis: ^9.2.0
   googleapis_auth: ^1.2.0
   intl: ^0.17.0
-  material_design_icons_flutter: ^5.0.6295
+  material_design_icons_flutter: ^6.0.7096
   url_launcher: ^6.0.13
 
 dev_dependencies:
-  bloc_test: ^8.5.0
+  bloc_test: ^9.1.0
   flutter_test:
     sdk: flutter
-  mocktail: ^0.2.0
-  very_good_analysis: ^2.4.0
+  mocktail: ^0.3.0
+  very_good_analysis: ^3.1.0
 
 flutter:
   uses-material-design: true
